refactor[#186] : constants 에 있던 스타일 파일setting.styles.ts 로 변경
