+++ src/features/setting/ui/AccountManagement.tsx
@@ -1,6 +1,6 @@
 import { Link } from "react-router-dom";
 import { ArrowRightIcon } from "@/shared/ui/icon";
-import { settingClassName } from "../constants";
+import { settingClassName } from "./setting.styles";
 
 // 계정 관리 하는 컴포넌트
 export const AccountManagement = () => (
